removed regex_brl and z_brl from link libraries
they should be contained in REGEX_STATIC_LIBRARIES and
ZLIB_STATIC_LIBRARIES respectively

If the regex and/or z library was compiled from the BRL-CAD sources,
these CMake variables should contain the names of the static library
versions. If they were taken from system, the variables should contain
the names of the system libraries (usually shared). If this is not the
case, the FindBRLCAD.cmake script had to be fixed.

...and some minor stylistic changes...